[COMMON][REVIEWS] 리뷰 기능 통합 정리
 바로 이해 되던 부분들은 일단 수정중에 있었습니다.
 컴파일 오류 난 부분들로 찾아가며 하던 중입니다.

+++ common/src/main/java/com/project/hiptour/common/reviews/controller/ReviewController.java
@@ -31,7 +31,7 @@ public class ReviewController {
     @Operation(summary = "리뷰 목록 조회", description = "장소 ID(placeId)에 해당하는 리뷰 offset, limit 나눠 조회")
     @GetMapping("/place/{placeId}")
     public List<ReviewListResponseDto> getReviewsByPlaceId(
-            @Parameter(description = "장소 ID") @PathVariable Long placeId,
+            @Parameter(description = "장소 ID") @PathVariable int placeId,
             @Parameter(description = "처음 가져올 개수") @RequestParam int offset,
             @Parameter(description = "가져올 개수") @RequestParam int limit
     ) {
