+++ common/src/main/java/com/project/hiptour/common/reviews/service/ReviewService.java
@@ -6,6 +6,6 @@ import com.project.hiptour.common.reviews.dto.ReviewPinRequestDto;
 import java.util.List;
 
 public interface ReviewService {
-    List<ReviewListResponseDto> getReviewsByPlaceId(Long placeId, int offset, int limit);
+    List<ReviewListResponseDto> getReviewsByPlaceId(int placeId, int offset, int limit);
     void pinReview(Long reviewId, ReviewPinRequestDto requestDto);
 }
